Extract cache feed image data use case into the new  test case class
